calculating suite result for nested specs - added SuccessRate, ExecutionTime, ExecutionStatus, SpecResults, PassedSpecCount, FailedSpecCount, SkippedSpecCount

+++ generator/generate.go
@@ -92,7 +92,7 @@ type buildError struct {
 type SuiteResult struct {
 	ProjectName            string       `json:"projectName"`
 	Timestamp              string       `json:"timestamp"`
-	SuccessRate            int          `json:"successRate"`
+	SuccessRate            float32      `json:"successRate"`
 	Environment            string       `json:"environment"`
 	Tags                   string       `json:"tags"`
 	ExecutionTime          int64        `json:"executionTime"`
